+++ src/plugins/piral-breadcrumbs/src/create.ts
@@ -1,6 +1,6 @@
 import * as ptr from 'path-to-regexp';
 import * as actions from './actions';
-import { buildName, PiralPlugin, Dict } from 'piral-core';
+import { buildName, PiralPlugin, Dict, appendItem } from 'piral-core';
 import { DefaultBreadbrumbItem, DefaultBreadcrumbsContainer } from './default';
 import { PiletBreadcrumbsApi, BreadcrumbSettings, BreadcrumbRegistration } from './types';
 
@@ -62,7 +62,7 @@ export function createBreadcrumbsApi(config: DashboardConfig = {}): PiralPlugin<
       },
     }));
 
-    return (_, target) => {
+    return (api, target) => {
       const pilet = target.name;
       let next = 0;
 
@@ -79,6 +79,7 @@ export function createBreadcrumbsApi(config: DashboardConfig = {}): PiralPlugin<
             matcher: getMatcher(settings),
             settings,
           });
+          return () => api.unregisterBreadcrumb(name);
         },
         unregisterBreadcrumb(name) {
           const id = buildName(pilet, name);